+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## Unreleased
+
+### Added
+
+- `AltPayType` reference with `DigitalRubImage` and `DigitalRubLink` constants
+- `AltPayRequestBuilder` for `POST /payments/altpay/pay` endpoint
+
 ## v1.10.0
 
 ### Added